:muscle: Move `executor.ts` into `git` module
