+++ denops/gin/feat/edit/read.ts
@@ -10,7 +10,7 @@ import {
 import {
   parse as parseBufname,
 } from "https://deno.land/x/denops_std@v4.0.0/bufname/mod.ts";
-import { execute } from "../../core/executor.ts";
+import { execute } from "../../git/executor.ts";
 import { formatTreeish } from "./util.ts";
 
 export async function read(
